[PLAYER-5446] Move BottomOverlay and its children to shared dir
