+++ sdk/react/src/BridgeListener.js
@@ -3,7 +3,7 @@ import { Platform } from 'react-native';
 
 import { CONTENT_TYPES, OVERLAY_TYPES, SCREEN_TYPES } from './constants';
 import Log from './lib/log';
-import { parseInputArray as parseMarkersInputArray } from './markers/lib';
+import { parseInputArray as parseMarkersInputArray } from './lib/markers';
 
 /**
  * The OoyalaSkinBridgeListener handles all of the listening of Player events from the Bridge.
