+++ sdk/react/src/shared/BottomOverlay/BottomOverlay.styles.js
