+++ sdk/react/src/shared/BottomOverlay/ControlBar/ControlBar.styles.json
