+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/IconMarker/IconMarker.test.js
