+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/IconMarker/index.js
