+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/Marker.js
@@ -2,9 +2,9 @@
 
 import * as React from 'react';
 
-import IconMarker from '../IconMarker';
-import TextMarker from '../TextMarker';
-import type { Marker as MarkerType } from '../../types/Marker';
+import IconMarker from './IconMarker';
+import TextMarker from './TextMarker';
+import type { Marker as MarkerType } from '../../../../types/Marker';
 
 type Props = {
   accentColor?: ?string,
