+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/Marker.test.js
