+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/TextMarker/TextMarker.styles.js
@@ -2,7 +2,7 @@
 
 import { StyleSheet } from 'react-native';
 
-import { MARKERS_SIZES } from '../../../constants';
+import { MARKERS_SIZES } from '../../../../../constants';
 import commonMarkerStyle from '../styles/commonMarkerStyle';
 import triangleStyle from '../styles/triangleStyle';
