+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/TextMarker/TextMarker.test.js
