+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/TextMarker/index.js
