+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/Marker/styles/commonMarkerStyle.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { MARKERS_SIZES } from '../../../constants';
+import { MARKERS_SIZES } from '../../../../../constants';
 
 export default {
   backgroundColor: 'black',
