+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/MarkersContainer.js
@@ -4,9 +4,9 @@ import * as React from 'react';
 import { Animated } from 'react-native';
 import type { ViewStyleProp } from 'react-native/Libraries/StyleSheet/StyleSheet';
 
-import Marker from '../Marker';
+import Marker from './Marker';
 import styles from './MarkersContainer.styles';
-import type { Marker as MarkerType } from '../../types/Marker';
+import type { Marker as MarkerType } from '../../../types/Marker';
 
 type Props = {
   accentColor?: ?string,
