+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/MarkersContainer.styles.js
