+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/MarkersContainer.test.js
