+++ sdk/react/src/shared/BottomOverlay/MarkersContainer/index.js
