+++ sdk/react/src/shared/BottomOverlay/MarkersProgressBarOverlayContainer/MarkerProgressBarOverlay/MarkerProgressBarOverlay.styles.js
