+++ sdk/react/src/shared/BottomOverlay/MarkersProgressBarOverlayContainer/MarkerProgressBarOverlay/MarkerProgressBarOverlay.test.js
