+++ sdk/react/src/shared/BottomOverlay/MarkersProgressBarOverlayContainer/MarkersProgressBarOverlayContainer.styles.js
