+++ sdk/react/src/shared/BottomOverlay/MarkersProgressBarOverlayContainer/MarkersProgressBarOverlayContainer.test.js
