+++ sdk/react/src/shared/BottomOverlay/MarkersProgressBarOverlayContainer/index.js
