+++ sdk/react/src/shared/BottomOverlay/index.js
